requirements.txt
Django>=4.0,<=5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
httpx>=0.26,<1.0
drf-spectacular>=0.27,<1.0
python-decouple>=3.8,<4.0
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
